refactor(desktop): move the answer normalizer and journal carry-over out of ChatProvider
ChatProvider.swift is held under the agent-runtime convergence ceiling
(7,382 lines); the streaming fixes had pushed it over. The sentence-spacing
normalizer now lives in ChatProvider+AssistantTextNormalization.swift and
carryingLocalOnlyFields beside its only caller in
ChatProvider+JournalProjection.swift. No behaviour change; the file ends
smaller than it started.
