+++ desktop/macos/e2e/flows/chat-first-cohesive.yaml
@@ -50,6 +50,7 @@ covers:
   # and asserted hermetically in ChatStreamingRenderBudgetTests.
   - desktop/macos/Desktop/Sources/MainWindow/Components/ChatStreamingRenderProbe.swift
   - desktop/macos/Desktop/Sources/Chat/ChatTurnLifecycle.swift
+  - desktop/macos/Desktop/Sources/Providers/ChatProvider+AssistantTextNormalization.swift
   # The prose parse and measured height this file caches are exercised by the
   # same mounts the flow drives; the cache changes no visible behaviour.
   - desktop/macos/Desktop/Sources/MainWindow/Components/ChatProseRenderCache.swift
